requirements.txt
pymupdf
sentence-transformers
python-dotenv
openai
chromadb
langchain-community 
langchain-chroma
